requirements.txt
torch
torchvision
numpy
tqdm
pyyaml
nibabel
